requirements.txt
anthropic>=0.86.0
PyMuPDF>=1.27.0
python-pptx>=1.0.0
Pillow>=12.0.0
lxml>=6.0.0
